requirements.txt
pandas
flask
scikit-learn
seaborn
matplotlib
IPython
pyyaml